tesseract/requirements.txt
Python 3.7.16
pillow == 9.5.0
pyperclip == 1.8.2
numpy == 1.21.6
pyocr == 0.8.5

